Added bp_chunked method
